+++ eq_sale_stock/models/eq_stock_pack_operation.py
@@ -19,23 +19,9 @@
 #
 ##############################################################################
 
-{
-    'name': "Equitania Verkauf & Lager",
-    'license': 'AGPL-3',
-    'version': '1.0.16',
-    'category': 'sale_stock',
-    'description': """ Improved view for order items""",
-    'author': 'Equitania Software GmbH',
-    'summary': ' Erweiterung Verkauf / Lager',
-    'website': 'www.myodoo.de',
-    "depends" : ['base', 'base_setup', 'eq_stock', 'sales_team', 'sale_stock', 'eq_sale', 'eq_product'],
-    'data': [
-            'security/ir.model.access.csv',
-            'views/eq_open_sale_order_line_view.xml',
-            'views/eq_stock_picking_view.xml',
-            'views/eq_sql_function_on_update.xml',
-            'views/eq_stock_pack_operation_view.xml',
-             ],
-    "active": False,
-    "installable": True
-}
+from odoo import models, fields, api, _
+
+class eq_stock_pack_operation(models.Model):
+    _inherit = 'stock.pack.operation'
+
+    eq_description = fields.Text(string = "Description")