+++ eq_sale_stock/models/stock.py
@@ -20,7 +20,8 @@
 ##############################################################################
 
 from odoo import models, fields, api, _
-
+from odoo.exceptions import UserError
+from collections import namedtuple
 
 class eq_stock_picking_extension(models.Model):
     _inherit = 'stock.picking'
@@ -39,8 +40,6 @@ class eq_stock_picking_extension(models.Model):
         ir_module_obj = self.env['ir.module.module'].search([('name', '=', 'eq_sale_stock'),('write_date','<','2018-12-20 14:00:00')])
         if len(ir_module_obj) > 0:
             self._cr.execute('alter table stock_picking drop column eq_sale_order_id cascade')
-
-
 
     eq_sale_order_id = fields.Many2one('sale.order', 'SaleOrder', compute='_compute_sale_order_id', store=True)
     # eq_header_text = fields.Html(string="Header")
@@ -63,3 +62,27 @@ class eq_stock_picking_extension(models.Model):
             vals['eq_footer_text'] = found_sale_order.note
 
         return super(eq_stock_picking_extension, self).create(vals)
+
+    def _prepare_pack_ops(self, quants, forced_qties):
+        """Added eq_description to pack_op_vals"""
+        self.ensure_one()
+        res = super(eq_stock_picking_extension, self)._prepare_pack_ops(
+            quants, forced_qties,
+        )
+
+        # Create dictionary with key product_id and value description(name) from the corresponding move,
+        # after that put it in pack_op_vals depending on the product_id
+        product_description = dict()
+        picking_moves = self.move_lines.filtered(lambda move: move.state not in ('done', 'cancel'))
+        for move in picking_moves:
+            if move.name:
+                if move.product_id.id in product_description and product_description[move.product_id.id]:
+                    product_description[move.product_id.id] = product_description[move.product_id.id] + " " \
+                                                              + (move.name).strip()
+                else:
+                    product_description[move.product_id.id] = move.name
+
+        for pack_op_vals in res:
+            pack_op_vals['eq_description'] = product_description[pack_op_vals['product_id']]
+
+        return res